Refaktorer skjemaet i forskjellige filer
